requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-simple-history==3.8.0
gunicorn==20.1.0
djangorestframework_simplejwt==5.4.0

